reduce number of files and share lookup logic
