+++ Package/gulpfile.js
@@ -217,6 +217,8 @@ const copyCommonFiles = () => {
 
 const copySharedFiles = () => {
   return gulp.src('../Apps/Playground/node_modules/@babylonjs/react-native/shared/BabylonNative.h')
+    .pipe(gulp.src('../Apps/Playground/node_modules/@babylonjs/react-native/shared/XrContextHelper.h'))
+    .pipe(gulp.src('../Apps/Playground/node_modules/@babylonjs/react-native/shared/XrAnchorHelper.h'))
     .pipe(gulp.dest('Assembled/shared'));
 };
 
@@ -244,8 +246,7 @@ const copyAndroidFiles = async () => {
   });
 
   await new Promise(resolve => {
-    gulp.src('../Apps/Playground/node_modules/@babylonjs/react-native/android/include/*')
-      .pipe(gulp.src('../Apps/Playground/node_modules/@babylonjs/react-native/submodules/BabylonNative/Dependencies/xr/src/ARCore/Include/*'))
+    gulp.src('../Apps/Playground/node_modules/@babylonjs/react-native/submodules/BabylonNative/Dependencies/xr/src/ARCore/Include/*'))
       .pipe(gulp.dest('Assembled/android/include'))
       .on('end', resolve);
   })
@@ -395,6 +396,8 @@ const validate = async () => {
 Assembled/EngineHook.ts
 Assembled/shared
 Assembled/shared/BabylonNative.h
+Assembled/shared/XrContextHelper.h
+Assembled/shared/XrAnchorHelper.h
 Assembled/EngineView.tsx
 Assembled/ios
 Assembled/ios/BabylonNativeInterop.mm
@@ -440,8 +443,6 @@ Assembled/android
 Assembled/android/build.gradle
 Assembled/android/include
 Assembled/android/include/IXrContextARCore.h
-Assembled/android/include/XrContextHelperARCore.h
-Assembled/android/include/XrAnchorHelperARCore.h
 Assembled/android/src
 Assembled/android/src/main
 Assembled/android/src/main/AndroidManifest.xml
